split app and server into 2 separate files
https://coveord.atlassian.net/browse/CDX-236
