+++ packages/search-token-server/server.ts
@@ -1,29 +1,8 @@
-/* eslint-disable @typescript-eslint/no-namespace */
-import * as express from 'express';
-import {ensureTokenGenerated} from './middlewares/searchToken';
 import {config} from 'dotenv';
-import {errorHandler} from './middlewares/errorHandler';
-import {environmentCheck} from './middlewares/environmentCheck';
 config();
 
-const app = express();
+import app from './app';
 const port = process.env.SERVER_PORT || 3000;
-
-declare global {
-  namespace Express {
-    interface Request {
-      token: string;
-    }
-  }
-}
-
-app.use(express.json());
-
-app.get('/token', environmentCheck, ensureTokenGenerated, (req, res) => {
-  res.send({token: req.token});
-});
-
-app.use(errorHandler);
 
 app.listen(port, () => {
   console.log(`Listening at http://localhost:${port}`);
